requirements.txt
# Medical Imaging Framework Requirements
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
PyQt5>=5.15.9
pyqtgraph>=0.13.3
nibabel>=5.1.0
pydicom>=2.4.0
SimpleITK>=2.2.1
monai>=1.2.0
tensorboard>=2.13.0
tqdm>=4.65.0
pyyaml>=6.0
pillow>=10.0.0
opencv-python>=4.8.0
albumentations>=1.3.1
scipy>=1.10.0
pandas>=2.0.0
